+++ src/components/ButtonRow/index.js
@@ -20,7 +20,7 @@ const ButtonRowSide = (side, buttons, testSection) => {
     'flex': includesFullWidthButton,
     'flex--1': true,
     [`oui-button-row--${side}`]: side,
-    'highlight-react--oui': localStorage.getItem('show_ouireact') == 'true',
+    'highlight-react--oui': localStorage.getItem('show_ouireact') === 'true',
   });
 
   return (