+++ src/components/Radio/index.js
@@ -20,7 +20,7 @@ const Radio = (props) => {
 
   const classes = classNames({
     'flex--none': true,
-    'highlight-react--oui': localStorage.getItem('show_ouireact') == 'true',
+    'highlight-react--oui': localStorage.getItem('show_ouireact') === 'true',
   });
 
   return (