+++ src/components/Token/index.js
@@ -13,7 +13,7 @@ const Token = (props) => {
   const classes = classNames({
     'oui-token-wrap': true,
     'flex': true,
-    'highlight-react--oui': localStorage.getItem('show_ouireact') == 'true',
+    'highlight-react--oui': localStorage.getItem('show_ouireact') === 'true',
   });
   const tokenToolsClasses = classNames({
     'oui-token-tool': props.isDraggable,